Fix: removed profile command
Instead, changed it to message_users so now it simple and available for admin only command.
Updated requirements.txt
