+++ src/commands/__init__.py
@@ -2,7 +2,7 @@ from .crypto import crypto_command_handler
 from .currency import currency_command_handler
 from .feedback import feedback_handler, feedback_reply_handler
 from .help import help_command_handler
-from .profile import profile_conversation_handler
+from .message_users import profile_conversation_handler
 from .ru_losses import ru_losses_handler
 from .settings import settings_conversation_handler
 from .start import start_command_handler
